// src/types/auth.ts
export type LogResult = 'SUCCESS' | 'FAILED' | 'EXPIRED' | 'LOCKED' | 'UNAUTHORIZED';
// ...
export interface AuthLog {
  _id: string;
  timestamp: string;
  userEmail: string;
  role: string;
  event: string;
  factor: 'PASSWORD' | 'MOBILE_OTP' | 'EMAIL_OTP' | 'SESSION' | 'SYSTEM';
  action: string;
  result: LogResult;
  details: string;
  ipAddress: string;
  userAgent: string;
  sessionId?: string;
}